feat: refactor and rework device and wallet frontend
Now the wallet frontend is included with each wallet canister. Also refactored
most files in the frontend to be TypeScript, specialized the frontend to connect
to its own canister only, and fixed typings and changes from the backend.

On the wallet side itself, keep the device list across upgrades and also
authorize new devices as custodians.